% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_import.R
\name{ftp_ctd_import}
\alias{ftp_ctd_import}
\title{Get single file imported}
\usage{
ftp_ctd_import(path)
}
\arguments{
\item{path}{the path to the txt file (ASCII)}
}
\description{
Works strictly for qc casts paths
}
