% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_import.R
\name{get_qc_path}
\alias{get_qc_path}
\title{Get path name to ctd file}
\usage{
get_qc_path(code, proj, cur_files = list_qc_ctd(proj))
}
\arguments{
\item{code}{the cruise number}

\item{proj}{the project name}

\item{cur_files}{the current file list}
}
\description{
Get path name to ctd file
}
