% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_import.R
\name{list_qc_ctd}
\alias{list_qc_ctd}
\title{Read list of available CTD files}
\usage{
list_qc_ctd(proj)
}
\arguments{
\item{proj}{either 'bats', 'gf', 'bv', 'bloom' or 'hs'}
}
\description{
Gets the names of the QC ctd files for either bats or hydrostation
}
