% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_import.R
\name{read_bats_ctd}
\alias{read_bats_ctd}
\title{Read QC ctd files}
\usage{
read_bats_ctd(code, proj)
}
\arguments{
\item{code}{the cruise code/number}

\item{proj}{the project acronym 'bats', 'bval', 'bloom', 'gf', 'hs'}
}
\description{
Reads a single ctd file off the bats ftp
check out the file information here.
}
